fix: try reconstructing path with /Fo and filename (#4)
Attempting to reconstruct the path using the /Fo argument only
considered the filename from the from the compile command entry in the
log. In some cases, however, a relative path is included in the compile
command entry and we may be able to reconstruct the absolute path from
it.

Consider:

    cl.exe /c /FoC:\\Path\\To\\ObjectFile ..\\src\\main.cpp

Prior to this commit, the components of the /Fo argument were joined
with just main.cpp, ignoring the ..\\src\\ components.

In addition to the original algorithm, the search also tries using
..\\src\\main.cpp with the components of the /Fo argument.